+++ code_generator/src/main/java/com/github/cowwoc/requirements/generator/ApiGenerator.java
@@ -52,6 +52,13 @@ public final class ApiGenerator
 	}
 
 	/**
+	 * Creates a new ApiGenerator.
+	 */
+	public ApiGenerator()
+	{
+	}
+
+	/**
 	 * The command-line entry point for this class.
 	 *
 	 * @param args the command-line arguments