+++ java/src/main/java/com/github/cowwoc/requirements/java/ListVerifier.java
@@ -7,6 +7,8 @@ package com.github.cowwoc.requirements.java;
 import com.github.cowwoc.requirements.java.extension.ExtensibleCollectionVerifier;
 import com.github.cowwoc.requirements.java.extension.ExtensibleObjectVerifier;
 
+import java.util.Comparator;
+
 /**
  * Verifies the requirements of a list.
  * <p>
@@ -17,4 +19,13 @@ import com.github.cowwoc.requirements.java.extension.ExtensibleObjectVerifier;
  */
 public interface ListVerifier<L, E> extends ExtensibleCollectionVerifier<ListVerifier<L, E>, L, E>
 {
+	/**
+	 * Ensures that the actual value is sorted.
+	 *
+	 * @param comparator the comparator that defines the order of the elements
+	 * @return the updated verifier
+	 * @throws IllegalArgumentException if the collection is not sorted
+	 * @see Comparator#naturalOrder()
+	 */
+	ListVerifier<L, E> isSorted(Comparator<E> comparator);
 }